tests: changed tests
